+++ lib/conventional_extensions/railtie.rb
@@ -1,0 +1,5 @@
+class ConventionalExtensions::Railtie < Rails::Railtie
+  initializer "conventional_extensions.ignore_extensions_directories" do
+    Rails.autoloaders.main.ignore "**/extensions/**.rb"
+  end
+end